+++ packages/api-fetch/src/types.ts
@@ -1,4 +1,5 @@
-export interface ApiFetchRequestProps {
+export interface APIFetchOptions extends RequestInit {
+	// Override headers, we only accept it as an object due to the `nonce` middleware
 	headers?: Record< string, string >;
 	path?: string;
 	url?: string;
@@ -9,10 +10,9 @@ export interface ApiFetchRequestProps {
 	data?: any;
 	namespace?: string;
 	endpoint?: string;
-	method?: string;
 }
 
-export type ApiFetchMiddleware = (
-	options: ApiFetchRequestProps,
-	next: ( nextOptions: ApiFetchRequestProps ) => Promise< any >
+export type APIFetchMiddleware = (
+	options: APIFetchOptions,
+	next: ( nextOptions: APIFetchOptions ) => Promise< any >
 ) => Promise< any >;