+++ packages/customize-widgets/src/components/sidebar-block-editor/index.js
@@ -1,11 +1,12 @@
 /**
  * WordPress dependencies
  */
-import { useReducer, createPortal } from '@wordpress/element';
+import { createPortal, useMemo } from '@wordpress/element';
 import {
 	BlockEditorProvider,
 	BlockList,
 	BlockSelectionClearer,
+	BlockInspector,
 	ObserveTyping,
 	WritingFlow,
 	BlockEditorKeyboardShortcuts,
@@ -20,39 +21,22 @@ import {
 /**
  * Internal dependencies
  */
-import Inspector, { BlockInspectorButton } from '../inspector';
+import BlockInspectorButton from '../block-inspector-button';
+import Header from '../header';
 import useSidebarBlockEditor from './use-sidebar-block-editor';
+import useInserter from '../inserter/use-inserter';
 
-const inspectorOpenStateReducer = ( state, action ) => {
-	switch ( action ) {
-		case 'OPEN':
-			return {
-				open: true,
-				busy: true,
-			};
-		case 'TRANSITION_END':
-			return {
-				...state,
-				busy: false,
-			};
-		case 'CLOSE':
-			return {
-				open: false,
-				busy: true,
-			};
-		default:
-			throw new Error( 'Unexpected action' );
-	}
-};
-
-export default function SidebarBlockEditor( { sidebar } ) {
+export default function SidebarBlockEditor( { sidebar, inserter, inspector } ) {
 	const [ blocks, onInput, onChange ] = useSidebarBlockEditor( sidebar );
-	const [
-		{ open: isInspectorOpened, busy: isInspectorAnimating },
-		setInspectorOpenState,
-	] = useReducer( inspectorOpenStateReducer, { open: false, busy: false } );
 	const parentContainer = document.getElementById(
 		'customize-theme-controls'
+	);
+	const [ isInserterOpened, setIsInserterOpened ] = useInserter( inserter );
+	const settings = useMemo(
+		() => ( {
+			__experimentalSetIsInserterOpened: setIsInserterOpened,
+		} ),
+		[]
 	);
 
 	return (
@@ -60,45 +44,46 @@ export default function SidebarBlockEditor( { sidebar } ) {
 			<BlockEditorKeyboardShortcuts.Register />
 			<SlotFillProvider>
 				<DropZoneProvider>
-					<div hidden={ isInspectorOpened && ! isInspectorAnimating }>
-						<BlockEditorProvider
-							value={ blocks }
-							onInput={ onInput }
-							onChange={ onChange }
-							useSubRegistry={ false }
-						>
-							<BlockEditorKeyboardShortcuts />
+					<BlockEditorProvider
+						value={ blocks }
+						onInput={ onInput }
+						onChange={ onChange }
+						settings={ settings }
+						useSubRegistry={ false }
+					>
+						<BlockEditorKeyboardShortcuts />
 
-							<BlockSelectionClearer>
-								<WritingFlow>
-									<ObserveTyping>
-										<BlockList />
-									</ObserveTyping>
-								</WritingFlow>
-							</BlockSelectionClearer>
-						</BlockEditorProvider>
+						<Header
+							inserter={ inserter }
+							isInserterOpened={ isInserterOpened }
+							setIsInserterOpened={ setIsInserterOpened }
+						/>
 
-						<Popover.Slot name="block-toolbar" />
-					</div>
+						<BlockSelectionClearer>
+							<WritingFlow>
+								<ObserveTyping>
+									<BlockList />
+								</ObserveTyping>
+							</WritingFlow>
+						</BlockSelectionClearer>
+					</BlockEditorProvider>
+
+					<Popover.Slot name="block-toolbar" />
 
 					{ createPortal(
-						<Inspector
-							isOpened={ isInspectorOpened }
-							isAnimating={ isInspectorAnimating }
-							setInspectorOpenState={ setInspectorOpenState }
-						/>,
-						parentContainer
+						// This is a temporary hack to prevent button component inside <BlockInspector>
+						// from submitting form when type="button" is not specified.
+						<form onSubmit={ ( event ) => event.preventDefault() }>
+							<BlockInspector />
+						</form>,
+						inspector.contentContainer[ 0 ]
 					) }
 
 					<__experimentalBlockSettingsMenuFirstItem>
 						{ ( { onClose } ) => (
 							<BlockInspectorButton
-								onClick={ () => {
-									// Open the inspector,
-									setInspectorOpenState( 'OPEN' );
-									// Then close the dropdown menu.
-									onClose();
-								} }
+								inspector={ inspector }
+								closeMenu={ onClose }
 							/>
 						) }
 					</__experimentalBlockSettingsMenuFirstItem>