+++ packages/create-block/CHANGELOG.md
@@ -4,7 +4,17 @@
 
 ### Enhancement
 
+-   Updated `.editorconfig` template files to work with automatic file formatting ([#30794](https://github.com/WordPress/gutenberg/pull/30794)).
+
+## 2.2.0 (2021-04-06)
+
+### Enhancement
+
 -   Scaffolded plugin requires WordPress 5.7 now ([#29757](https://github.com/WordPress/gutenberg/pull/29757)).
+
+### New Features
+
+-   Add new `theme` category to select for the block type ([#30089](https://github.com/WordPress/gutenberg/pull/30089)).
 
 ## 2.1.0 (2021-03-17)